Remove RiffRaff API
