--- riff-raff/app/views/api/form.scala.html
@@ -1,18 +1,0 @@
-@import conf.Config
-@(config: Config, menu: Menu)(apiForm: Form[String])(implicit request: Security.AuthenticatedRequest[AnyContent, com.gu.googleauth.UserIdentity], messages: Messages)
-@import helper.CSRF
-
-@main("Create new API key", request) {
-
-    <h2>Create new API key</h2>
-    <hr/>
-
-    @b3.vertical.form(action=routes.Api.createKey) { implicit fc =>
-        @CSRF.formField
-        @b3.text(apiForm("application"), Symbol("_label") -> "Application that will use this key")
-        <div class="actions">
-            <button name="action" type="submit" value="save" class="btn btn-primary">Create</button> or
-            <a href="@routes.Api.listKeys" class="btn btn-danger">Cancel</a>
-        </div>
-    }
-}(config, menu)
