--- riff-raff/app/views/api/list.scala.html
@@ -1,46 +1,0 @@
-@import conf.Config
-@(config: Config, menu: Menu)(implicit request: Security.AuthenticatedRequest[AnyContent, com.gu.googleauth.UserIdentity], keyList: Iterable[controllers.ApiKey])
-@import helper.CSRF
-
-@main("API keys", request) {
-    <h2>API keys</h2>
-    <hr/>
-    <p><a class="btn btn-primary" href="@routes.Api.createKeyForm"><i class="glyphicon glyphicon-plus glyphicon-white"></i> Create new API key</a></p>
-    <div class="content">
-        @if(keyList.isEmpty) {
-            <div class="alert alert-warning"><strong>No API keys.</strong></div>
-        } else {    
-            <table class="table table-condensed">
-                <thead>
-                    <tr>
-                        <th>Date created</th>
-                        <th>Created by</th>
-                        <th>Application</th>
-                        <th>Key</th>
-                        <th>Last used</th>
-                        <th>Times used</th>
-                        <th></th>
-                    </tr>
-                </thead>
-                <tbody>
-                @keyList.map { key =>
-                    <tr>
-                        <td>@utils.DateFormats.Short.print(key.created)</td>
-                        <td><span class="label label-default">@key.issuedBy</span></td>
-                        <td>@key.application</td>
-                        <td>@key.key</td>
-                        <td>@key.lastUsed.map(utils.DateFormats.Short.print).getOrElse("never")</td>
-                        <td>@key.totalCalls</td>
-                        <td>
-                        @helper.form(routes.Api.delete, Symbol("class") -> "modal-form") {
-                            @CSRF.formField
-                            <button class="btn btn-xs btn-danger" name="key" value="@key.key"><i class="glyphicon glyphicon-trash glyphicon glyphicon-white"></i> Delete</button>
-                        }
-                        </td>
-                    </tr>
-                }
-                </tbody>
-            </table>
-        }
-    </div>
-}(config, menu)
